networking and parsing json
We create a model for our Movie class that extrapolated data from JSONObjects that consist of the properties listed by the model.

We instantiate AsyncHttpClient in order to grab data from our API, so long as we add it as a dependency. We pass in the result to our model, so we can retrieve movies with the listed properties.

We practice logging with tags.

We practice method signatures in accordance to throwing JSONExceptions.